+++ CHANGES.rst
@@ -5,6 +5,12 @@ Changelog
 1.2.43 (unreleased)
 -------------------
 
+- WEB-4098 : Add affiche scale for section on table view display and one element by lot 
+  [thomlamb]
+
+- WEB-4098 : Add new profile for imio.smartweb.orientation to section files
+  [thomlamb]
+
 - SUP-35100 : Fix sitemap.xml.gz generation. When a "main" rest view was removed, continue to build sitemap for others rest views
   [boulch]